refactor: gender enum
