--- backend/src/user/user.enum.ts
@@ -1,11 +1,0 @@
-export enum UserRole {
-  SUPER_ADMIN = 'super_admin',
-  ARTIST_MANAGER = 'artist_manager',
-  ARTIST = 'artist',
-}
-
-export enum Gender {
-  MALE = 'male',
-  FEMALE = 'female',
-  OTHERS = 'other',
-}
